Kumaraswamy distribution maximum likelihood estimation

Uses Newton-Raphson to estimate the parameters of the Kumaraswamy distribution.

Usage

mlkumar(x, na.rm = FALSE, ...)

Arguments

x

a (non-empty) numeric vector of data values.

na.rm

logical. Should missing values be removed?

...

a0 is an optional starting value for the a parameter. rel.tol is the relative accuracy requested, defaults to .Machine$double.eps^0.25. iterlim is a positive integer specifying the maximum number of iterations to be performed before the program is terminated (defaults to 100).

Value

mlkumar returns an object of class univariateML. This is a named numeric vector with maximum likelihood estimates for a and b and the following attributes:

model

The name of the model.

density

The density associated with the estimates.

logLik

The loglikelihood at the maximum.

support

The support of the density.

n

The number of observations.

call

The call as captured my match.call

Details

For the density function of the Kumaraswamy distribution see Kumaraswamy.

References

Jones, M. C. "Kumaraswamy's distribution: A beta-type distribution with some tractability advantages." Statistical Methodology 6.1 (2009): 70-81.

Kumaraswamy, Ponnambalam. "A generalized probability density function for double-bounded random processes." Journal of Hydrology 46.1-2 (1980): 79-88.

See also

Examples

AIC(mlkumar(USArrests$Rape / 100))
#> [1] -96.06926