Power distribution maximum likelihood estimation

The maximum likelihood estimate of alpha is the maximum of x + epsilon (see the details) and the maximum likelihood estimate of beta is 1/(log(alpha)-mean(log(x))).

Usage

mlpower(x, na.rm = FALSE, ...)

Arguments

x

a (non-empty) numeric vector of data values.

na.rm

logical. Should missing values be removed?

...

epsilon is a positive number added to max(x) as an to the maximum likelihood. Defaults to .Machine$double.eps^0.5.

Value

mlpower returns an object of class univariateML. This is a named numeric vector with maximum likelihood estimates for alpha and beta and the following attributes:

model

The name of the model.

density

The density associated with the estimates.

logLik

The loglikelihood at the maximum.

support

The support of the density.

n

The number of observations.

call

The call as captured my match.call

Details

For the density function of the power distribution see PowerDist. The maximum likelihood estimator of alpha does not exist, strictly speaking. This is because x is supported c(0, alpha) with an open endpoint on alpha in the extraDistr implementation of dpower. If the endpoint was closed, max(x) would have been the maximum likelihood estimator. To overcome this problem, we add a possibly user specified epsilon to max(x).

References

Arslan, G. "A new characterization of the power distribution." Journal of Computational and Applied Mathematics 260 (2014): 99-102.

See also

Examples

mlpower(precip)
#> Maximum likelihood estimates for the PowerDist model 
#>  alpha    beta  
#> 67.000   1.312