Weibull distribution maximum likelihood estimation

Uses Newton-Raphson to estimate the parameters of the Weibull distribution.

Usage

mlweibull(x, na.rm = FALSE, ...)

Arguments

x

a (non-empty) numeric vector of data values.

na.rm

logical. Should missing values be removed?

...

shape0 is an optional starting value for the shape parameter. rel.tol is the relative accuracy requested, defaults to .Machine$double.eps^0.25. iterlim is a positive integer specifying the maximum number of iterations to be performed before the program is terminated (defaults to 100).

Value

mlweibull returns an object of class univariateML. This is a named numeric vector with maximum likelihood estimates for shape and scale and the following attributes:

model

The name of the model.

density

The density associated with the estimates.

logLik

The loglikelihood at the maximum.

support

The support of the density.

n

The number of observations.

call

The call as captured my match.call

Details

For the density function of the Weibull distribution see Weibull.

References

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, Volume 1, Chapter 21. Wiley, New York.

See also

Examples

BIC(mlweibull(precip))
#> [1] 573.3096