Fit multiple models and select the best fit

Selects the best model by log-likelihood, AIC, or BIC.

Usage

model_select(
  x,
  models = univariateML_models,
  criterion = c("aic", "bic", "loglik"),
  na.rm = FALSE,
  rtn_class = "univariateML",
  ...
)

Arguments

x

a (non-empty) numeric vector of data values.

models

a character vector containing the distribution models to select from; see print(univariateML_models).

criterion

the model selection criterion. Must be one of "aic", "bic", and "loglik". Defaults to "aic".

na.rm

logical. Should missing values be removed?

rtn_class

character length 1. "univariateML" (default) if the function should return the single best model; "data.frame" if a tibble data frame of all results should be returned, sorted by decreasing model performance.

...

unused.

Value

The return value depends on the rtn_class argument. For rtn_class = "univariateML" (default), model_select returns an object of class #' univariateML. This is a named numeric vector with maximum likelihood estimates for the parameters of the best fitting model and the following attributes:

model

The name of the model.

density

The density associated with the estimates.

logLik

The loglikelihood at the maximum.

support

The support of the density.

n

The number of observations.

call

The call as captured my match.call

For rtn_class = "data.frame", model_select returns a tibble data frame with the following columns:

model

The name of the model.

d_loglik, d_aic, d_bic

See loglik, aic, bic.

loglik, aic, bic

The negative loglikelihood at the maximum, the AIC, and the BIC, respectively. The minimum of each of these is noted and then subtracted from each value to give their delta versions d_loglik, d_aic, d_bic

. So, the model with the lowest AIC will have d_aic of 0; the d_aic of all the other models shows how much higher their AICs are from the minimum. The same goes with d_loglik and d_bic.

ml

The internal code name for the model.

univariateML

The univariateML object for the model. This is identical to the value returned for rtn_class = "univariateML"; for rtn_class = "data.frame", this object is returned for all tested models, not just the best one.

See also

Examples

# Select among all possible models.
model_select(precip)
#> Maximum likelihood estimates for the Skew Normal model 
#>    mean       sd       xi  
#> 34.6957  13.5471   0.8088  

# View possible models to fit.
print(univariateML_models)
#>  [1] "beta"       "betapr"     "cauchy"     "exp"        "gamma"     
#>  [6] "ged"        "gumbel"     "invgamma"   "invgauss"   "invweibull"
#> [11] "kumar"      "laplace"    "lgamma"     "llogis"     "lnorm"     
#> [16] "logis"      "logitnorm"  "lomax"      "naka"       "norm"      
#> [21] "pareto"     "pois"       "power"      "rayleigh"   "sged"      
#> [26] "snorm"      "sstd"       "std"        "unif"       "weibull"   

# Try out only gamma, Weibull, and exponential.
model_select(precip, c("gamma", "weibull", "exp"))
#> Maximum likelihood estimates for the Weibull model 
#>  shape   scale  
#>  2.829  39.084