Abalone data

Physical measurements of 4177 abalones, a species of sea snail.

Usage

abalone

Format

A tibble with 4,177 observations and 9 variables:

sex

Sex of the abalone, F is female, M male, and I infant.

length

Longest shell measurement.

diameter

Diameter perpendicular to length.

height

Height with with meat in shell.

whole_weight

Grams whole abalone.

shucked_weight

Grams weight of meat.

viscera_weight

Grams gut weight (after bleeding).

shell_weight

Grams after being dried.

rings

+1.5 gives the age in years.

Source

Dua, D. and Graff, C. (2019). UCI Machine Learning Repository https://archive.ics.uci.edu/ml/. Irvine, CA: University of California, School of Information and Computer Science.

Details

See the web page https://archive.ics.uci.edu/ml/datasets/Abalone for more information about the data set.

References

Ko, V., Hjort, N. L., & Hobaek Haff, I. (2019). Focused information criteria for copulas. Scandinavian Journal of Statistics.

Examples

abalone
#> # A tibble: 4,177 × 9
#>    sex   length diameter height whole_weight shucked_weight viscera_weight
#>    <fct>  <dbl>    <dbl>  <dbl>        <dbl>          <dbl>          <dbl>
#>  1 M      0.455    0.365  0.095        0.514         0.224          0.101 
#>  2 M      0.35     0.265  0.09         0.226         0.0995         0.0485
#>  3 F      0.53     0.42   0.135        0.677         0.256          0.142 
#>  4 M      0.44     0.365  0.125        0.516         0.216          0.114 
#>  5 I      0.33     0.255  0.08         0.205         0.0895         0.0395
#>  6 I      0.425    0.3    0.095        0.352         0.141          0.0775
#>  7 F      0.53     0.415  0.15         0.778         0.237          0.142 
#>  8 F      0.545    0.425  0.125        0.768         0.294          0.150 
#>  9 M      0.475    0.37   0.125        0.509         0.216          0.112 
#> 10 F      0.55     0.44   0.15         0.894         0.314          0.151 
#> # … with 4,167 more rows, and 2 more variables: shell_weight <dbl>, rings <int>