Confidence Intervals for Maximum Likelihood Estimates

Computes a confidence interval for one or more parameters in a unvariateML object.

Usage

# S3 method for class 'univariateML'
confint(object, parm = NULL, level = 0.95, Nreps = 1000, ...)

Arguments

object

An object of class univariateML.

parm

Vector of strings; the parameters to calculate a confidence interval for. Each parameter must be a member of names(object).

level

The confidence level.

Nreps

Number of bootstrap iterations. Passed to bootstrapml().

...

Additional arguments passed to bootstrapml().

Value

A matrix or vector with columns giving lower and upper confidence limits for each parameter in parm.

Details

confint.univariateML is a wrapper for bootstrapml() that computes confidence intervals for the main parameters of object. The main parameters of object are the members of names(object). For instance, the main parameters of an object obtained from mlnorm are mean and sd. The confidence intervals are parametric bootstrap percentile intervals with limits (1-level)/2 and 1 - (1-level).

See also

Examples

object <- mlinvgauss(airquality$Wind)
confint(object) # 95% confidence interval for mean and shape
#>            2.5%    97.5%
#> mean   9.282952 10.63857
#> shape 44.413492 71.16450
confint(object, "mean") # 95% confidence interval for the mean parameter
#>      2.5%     97.5% 
#>  9.293843 10.585785 
# confint(object, "variance") # Fails since 'variance isn't a main parameter.